change partner type to uuid and move partner module into users
