+++ nest-cli.json
@@ -2,7 +2,8 @@
   "collection": "@nestjs/schematics",
   "sourceRoot": "src",
   "compilerOptions": {
-    "assets": ["**/*.ejs"],
-    "watchAssets": true
+    "assets": ["mail/templates/**/*"],
+    "watchAssets": true,
+    "outDir": "./dist/src"
   }
 }
