+++ package.json
@@ -32,7 +32,7 @@
     "@willsoto/nestjs-objection": "^5.0.0",
     "bcrypt": "^5.0.1",
     "class-validator": "^0.13.1",
-    "ejs": "^3.1.6",
+    "handlebars": "^4.7.7",
     "helmet": "^4.6.0",
     "knex": "^0.21.19",
     "nest-winston": "^1.5.0",
