+++ src/app.module.ts
@@ -8,10 +8,9 @@ import { APP_GUARD } from '@nestjs/core';
 import { JwtAuthGuard } from './auth/guards/jwt-auth.guard';
 import { RolesGuard } from './auth/guards/roles.guard';
 import { MailModule } from './mail/mail.module';
-import { PartnerModule } from './partner/partner.module';
 
 @Module({
-  imports: [DatabaseModule, UsersModule, AuthModule, MailModule, PartnerModule],
+  imports: [DatabaseModule, UsersModule, AuthModule, MailModule],
   controllers: [AppController],
   providers: [
     AppService,
