+++ src/db/seeds/20210515074708_insert_data_into_users.ts
@@ -11,7 +11,7 @@ export async function seed(knex: Knex): Promise<void> {
       password: '$2b$10$Z5Wt2IbY1KEQbuEunfFlguWUIjsH4P.nlhpvzryrTatn8Oc9dildu',
       role: 'admin',
       id: '9bdac0a0-a682-43bc-865a-32b983f94f0a',
-      partner: '',
+      partner: null,
       status: 'single',
       requestedAt: null,
     },
@@ -21,7 +21,7 @@ export async function seed(knex: Knex): Promise<void> {
       password: '$2b$10$7hkF9rSehP8E.2Aa3i0uFufBNaHp0W6MUwKUaHe3cI2/5.ZvzEQ5C',
       role: 'user',
       id: 'e259b4f7-00c8-4ce2-8dd0-c54e26b97dd5',
-      partner: '',
+      partner: null,
       status: 'single',
       requestedAt: null,
     },
@@ -31,7 +31,7 @@ export async function seed(knex: Knex): Promise<void> {
       password: '$2b$10$i1M9HOtIbArQGPXp3gJiJudMPWQPQoiW7PLRIP6w9otEBrIRjAiNe',
       role: 'user',
       id: 'e7bee3e9-a5d2-4378-860e-3e7a85803e11',
-      partner: '',
+      partner: null,
       status: 'single',
       requestedAt: null,
     },
@@ -41,7 +41,7 @@ export async function seed(knex: Knex): Promise<void> {
       password: '$2b$10$WVrEmVYFpDOMBHgVhTTwtOJ37O9Hu9EbRdEil2JnAeBl/L3BC1yde',
       role: 'user',
       id: 'fffd3b1e-bef8-404e-a15d-2586f79da920',
-      partner: '',
+      partner: null,
       status: 'single',
       requestedAt: null,
     },
