--- src/partner/partner.controller.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { PartnerController } from './partner.controller';
-
-describe('PartnerController', () => {
-  let controller: PartnerController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [PartnerController],
-    }).compile();
-
-    controller = module.get<PartnerController>(PartnerController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
