--- src/partner/partner.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { PartnerService } from './partner.service';
-
-describe('PartnerService', () => {
-  let service: PartnerService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [PartnerService],
-    }).compile();
-
-    service = module.get<PartnerService>(PartnerService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
